refactor: extract SIR optimizer crate
