+++ crates/celox-sir-opt/Cargo.toml
@@ -8,8 +8,15 @@ description = "Backend-independent SIR optimization policy and passes for Celox"
 edition.workspace = true
 
 [dependencies]
-fxhash = { workspace = true }
+bit-set             = { workspace = true }
+celox-analysis      = { path = "../celox-analysis" }
+celox-design        = { path = "../celox-design" }
+celox-sir           = { path = "../celox-sir" }
+celox-state-layout  = { path = "../celox-state-layout" }
+fxhash              = { workspace = true }
+itertools           = { workspace = true }
+num-bigint          = "0.4"
+num-traits          = { workspace = true }
 
 [lints]
 workspace = true
-
