+++ crates/celox-sir-opt/src/coalescing/block_opt.rs
