+++ crates/celox-sir-opt/src/coalescing/commit_ops.rs
