+++ crates/celox-sir-opt/src/coalescing/control_region_feasibility.rs
