+++ crates/celox-sir-opt/src/coalescing/cost_model.rs
