+++ crates/celox-sir-opt/src/coalescing/dead_working_stores.rs
