+++ crates/celox-sir-opt/src/coalescing/pass_bit_extract_peephole.rs
