+++ crates/celox-sir-opt/src/coalescing/pass_branchify_mux.rs
