+++ crates/celox-sir-opt/src/coalescing/pass_coalesce_stores.rs
