+++ crates/celox-sir-opt/src/coalescing/pass_commit_sinking.rs
