+++ crates/celox-sir-opt/src/coalescing/pass_concat_folding.rs
