+++ crates/celox-sir-opt/src/coalescing/pass_control_flow_simplify.rs
