+++ crates/celox-sir-opt/src/coalescing/pass_dead_code_elimination.rs
