+++ crates/celox-sir-opt/src/coalescing/pass_effect_case_dispatch.rs
