+++ crates/celox-sir-opt/src/coalescing/pass_eliminate_dead_working_stores.rs
