+++ crates/celox-sir-opt/src/coalescing/pass_eliminate_working_round_trip.rs
@@ -13,8 +13,7 @@ use crate::ir::*;
 ///
 /// `eu_boundary_blocks`: block IDs that start a new original EU (after SIR merge).
 /// Empty for pre-merge (single EU) — all variables are trivially independent.
-#[cfg_attr(not(target_arch = "x86_64"), allow(dead_code))]
-pub(crate) fn eliminate_working_round_trip(
+pub fn eliminate_working_round_trip(
     eu: &mut ExecutionUnit<RegionedAbsoluteAddr>,
     eu_boundary_blocks: &[BlockId],
 ) {
