+++ crates/celox-sir-opt/src/coalescing/pass_gvn.rs
