+++ crates/celox-sir-opt/src/coalescing/pass_hoist_common_branch_loads.rs
