+++ crates/celox-sir-opt/src/coalescing/pass_identity_store_bypass.rs
