+++ crates/celox-sir-opt/src/coalescing/pass_indexed_store_recovery.rs
