+++ crates/celox-sir-opt/src/coalescing/pass_inline_commit_forwarding.rs
