+++ crates/celox-sir-opt/src/coalescing/pass_loop_idiom.rs
