+++ crates/celox-sir-opt/src/coalescing/pass_manager.rs
