+++ crates/celox-sir-opt/src/coalescing/pass_masked_array_any.rs
