+++ crates/celox-sir-opt/src/coalescing/pass_optimize_blocks.rs
