+++ crates/celox-sir-opt/src/coalescing/pass_pack_concat_phi.rs
