+++ crates/celox-sir-opt/src/coalescing/pass_packed_scatter_store.rs
