+++ crates/celox-sir-opt/src/coalescing/pass_phi_outcome_compression.rs
