+++ crates/celox-sir-opt/src/coalescing/pass_sparse_case_dispatch.rs
