+++ crates/celox-sir-opt/src/coalescing/pass_split_coalesced_stores.rs
