+++ crates/celox-sir-opt/src/coalescing/pass_split_wide_commits.rs
