+++ crates/celox-sir-opt/src/coalescing/pass_vectorize_concat.rs
@@ -15,9 +15,6 @@
 
 use super::pass_manager::ExecutionUnitPass;
 use super::shared::{def_reg, sir_value_to_u64};
-#[cfg(any(target_arch = "x86_64", test))]
-use crate::backend::MemoryLayout;
-#[cfg(any(target_arch = "x86_64", test))]
 use crate::ir::cfg::SirCfg;
 use crate::ir::*;
 use crate::optimizer::PassOptions;
@@ -150,14 +147,12 @@ struct PackedBitStoreSink {
 }
 
 #[derive(Clone, Copy)]
-#[cfg(any(target_arch = "x86_64", test))]
 enum PackedChainDefinition {
     Parameter(BlockId),
     Instruction(BlockId, usize),
 }
 
 #[derive(Clone)]
-#[cfg(any(target_arch = "x86_64", test))]
 struct PackedConditionalStoreChain {
     start: BlockId,
     endpoint: BlockId,
@@ -169,7 +164,6 @@ struct PackedConditionalStoreChain {
 }
 
 #[derive(Clone, Copy)]
-#[cfg(any(target_arch = "x86_64", test))]
 struct PackedConditionalStoreDiamond {
     predicate: RegisterId,
     destination: RegionedAbsoluteAddr,
@@ -180,7 +174,6 @@ struct PackedConditionalStoreDiamond {
     empty_arm: BlockId,
 }
 
-#[cfg(any(target_arch = "x86_64", test))]
 fn packed_chain_definitions(
     eu: &ExecutionUnit<RegionedAbsoluteAddr>,
 ) -> HashMap<RegisterId, PackedChainDefinition> {
@@ -201,7 +194,6 @@ fn packed_chain_definitions(
     definitions
 }
 
-#[cfg(any(target_arch = "x86_64", test))]
 fn packed_chain_value_dominates(
     definitions: &HashMap<RegisterId, PackedChainDefinition>,
     cfg: &SirCfg,
@@ -218,7 +210,6 @@ fn packed_chain_value_dominates(
     }
 }
 
-#[cfg(any(target_arch = "x86_64", test))]
 fn packed_chain_bit_constant(
     eu: &ExecutionUnit<RegionedAbsoluteAddr>,
     definitions: &HashMap<RegisterId, PackedChainDefinition>,
@@ -243,7 +234,6 @@ fn packed_chain_bit_constant(
     None
 }
 
-#[cfg(any(target_arch = "x86_64", test))]
 fn packed_chain_single_predecessor(cfg: &SirCfg, block: BlockId, expected: BlockId) -> bool {
     let Some(index) = cfg.block_index(block) else {
         return false;
@@ -251,7 +241,6 @@ fn packed_chain_single_predecessor(cfg: &SirCfg, block: BlockId, expected: Block
     cfg.predecessors[index].as_slice() == cfg.block_index(expected).as_slice()
 }
 
-#[cfg(any(target_arch = "x86_64", test))]
 fn packed_chain_merge_predecessors(
     cfg: &SirCfg,
     merge: BlockId,
@@ -272,7 +261,6 @@ fn packed_chain_merge_predecessors(
     actual == expected
 }
 
-#[cfg(any(target_arch = "x86_64", test))]
 fn packed_constant_store_arm(
     eu: &ExecutionUnit<RegionedAbsoluteAddr>,
     definitions: &HashMap<RegisterId, PackedChainDefinition>,
@@ -304,7 +292,6 @@ fn packed_constant_store_arm(
         .then_some((*destination, *offset, stored_value, *merge))
 }
 
-#[cfg(any(target_arch = "x86_64", test))]
 fn packed_empty_store_arm(
     eu: &ExecutionUnit<RegionedAbsoluteAddr>,
     arm: BlockId,
@@ -319,7 +306,6 @@ fn packed_empty_store_arm(
     arguments.is_empty().then_some(*merge)
 }
 
-#[cfg(any(target_arch = "x86_64", test))]
 fn packed_conditional_store_diamond(
     eu: &ExecutionUnit<RegionedAbsoluteAddr>,
     cfg: &SirCfg,
@@ -379,32 +365,6 @@ fn packed_conditional_store_diamond(
     })
 }
 
-#[cfg(any(target_arch = "x86_64", test))]
-fn packed_range_is_physically_contiguous(
-    layout: &MemoryLayout,
-    destination: RegionedAbsoluteAddr,
-    start: usize,
-    width: usize,
-) -> bool {
-    let Some((base_byte, base_bit)) = layout.regioned_static_byte_and_intra(&destination, start)
-    else {
-        return false;
-    };
-    (0..width).all(|bit| {
-        let Some(bit_offset) = start.checked_add(bit) else {
-            return false;
-        };
-        let Some((byte, intra)) = layout.regioned_static_byte_and_intra(&destination, bit_offset)
-        else {
-            return false;
-        };
-        let physical_bit = base_bit + bit;
-        byte == base_byte + i32::try_from(physical_bit / 8).unwrap_or(i32::MAX)
-            && intra == physical_bit % 8
-    })
-}
-
-#[cfg(any(target_arch = "x86_64", test))]
 fn plan_packed_conditional_store_chain(
     eu: &ExecutionUnit<RegionedAbsoluteAddr>,
     cfg: &SirCfg,
@@ -465,8 +425,7 @@ fn plan_packed_conditional_store_chain(
     })
 }
 
-#[cfg(any(target_arch = "x86_64", test))]
-fn collapse_packed_conditional_store_chains_with(
+pub(crate) fn collapse_packed_conditional_store_chains_with(
     eu: &mut ExecutionUnit<RegionedAbsoluteAddr>,
     physically_contiguous: impl Fn(RegionedAbsoluteAddr, usize, usize) -> bool,
 ) -> usize {
@@ -585,19 +544,6 @@ fn collapse_packed_conditional_store_chains_with(
     plans.len()
 }
 
-/// Collapse a serial CFG of conditional one-bit constant stores into one
-/// packed read/modify/write. The physical-layout proof prevents a semantic
-/// packed range from becoming an invalid cross-element native access.
-#[cfg(any(target_arch = "x86_64", test))]
-pub(super) fn collapse_native_packed_conditional_store_chains(
-    eu: &mut ExecutionUnit<RegionedAbsoluteAddr>,
-    layout: &MemoryLayout,
-) -> usize {
-    collapse_packed_conditional_store_chains_with(eu, |destination, start, width| {
-        packed_range_is_physically_contiguous(layout, destination, start, width)
-    })
-}
-
 /// Make a complete set of scalar one-bit publications consume the packed
 /// value which already represents those bits.
 ///
@@ -608,7 +554,6 @@ pub(super) fn collapse_native_packed_conditional_store_chains(
 /// preserves their observable order while giving the vectorizer one real
 /// packed sink. Native ISel may subsequently combine the exact Slice/Store
 /// pairs when the physical element layout permits it.
-#[cfg(target_arch = "x86_64")]
 pub(super) fn expose_packed_bit_store_sinks(eu: &mut ExecutionUnit<RegionedAbsoluteAddr>) -> bool {
     let mut next_register = eu
         .register_map
