+++ crates/celox-sir-opt/src/coalescing/pass_xor_chain_folding.rs
