+++ crates/celox-sir-opt/src/coalescing/placement_analysis.rs
