+++ crates/celox-sir-opt/src/coalescing/shared.rs
