+++ crates/celox-sir-opt/src/coalescing/sir_analysis.rs
