+++ crates/celox/src/parser.rs
@@ -198,196 +198,7 @@ pub(crate) fn verify_memory_offset_contract(
     program: &Program,
     eu: &crate::ir::ExecutionUnit<RegionedAbsoluteAddr>,
 ) -> Result<(), crate::ir::verify::SirVerifyError> {
-    use crate::ir::SIRInstruction;
-
-    for block in eu.blocks.values() {
-        for (index, inst) in block.instructions.iter().enumerate() {
-            let (addr, offset, width, operation, explicit_memory_copy) = match inst {
-                SIRInstruction::Load(_, addr, offset, width) => {
-                    (addr, offset, *width, "Load", false)
-                }
-                SIRInstruction::Store(addr, offset, width, _, _, _) => {
-                    (addr, offset, *width, "Store", false)
-                }
-                SIRInstruction::Commit(src, dst, offset, width, _) => {
-                    verify_memory_offset_for_addr(
-                        program,
-                        block.id,
-                        index,
-                        dst,
-                        offset,
-                        *width,
-                        "Commit destination",
-                        true,
-                    )?;
-                    (src, offset, *width, "Commit source", true)
-                }
-                _ => continue,
-            };
-            verify_memory_offset_for_addr(
-                program,
-                block.id,
-                index,
-                addr,
-                offset,
-                width,
-                operation,
-                explicit_memory_copy,
-            )?;
-        }
-    }
-    Ok(())
-}
-
-fn verify_memory_offset_for_addr(
-    program: &Program,
-    block: crate::ir::BlockId,
-    index: usize,
-    addr: &RegionedAbsoluteAddr,
-    offset: &crate::ir::SIROffset,
-    width: usize,
-    operation: &'static str,
-    explicit_memory_copy: bool,
-) -> Result<(), crate::ir::verify::SirVerifyError> {
-    use crate::ir::SIROffset;
-
-    let Some(info) = program.get_variable_info(&addr.absolute_addr()) else {
-        return Err(crate::ir::verify::SirVerifyError::instruction(
-            "MEMORY.ADDRESS_HAS_DECLARATION",
-            block,
-            index,
-            format!("no variable declaration for memory address {addr:?}"),
-        ));
-    };
-    let element_count = info
-        .array_dims
-        .iter()
-        .try_fold(1usize, |count, &dimension| count.checked_mul(dimension));
-    let declared_element_width = (!info.array_dims.is_empty())
-        .then_some(element_count)
-        .flatten()
-        .filter(|&count| count != 0 && info.width % count == 0)
-        .map(|count| info.width / count);
-
-    match offset {
-        SIROffset::Dynamic(_) if !info.array_dims.is_empty() => {
-            let absolute_addr = addr.absolute_addr();
-            return Err(crate::ir::verify::SirVerifyError::instruction(
-                "MEMORY.UNPACKED_OFFSET_IS_ELEMENT",
-                block,
-                index,
-                format!(
-                    "{operation} addresses unpacked array {} with dimensions {:?} by an arbitrary dynamic bit offset; preserve the element index as SIROffset::Element",
-                    program.get_path(&absolute_addr),
-                    info.array_dims,
-                ),
-            ));
-        }
-        SIROffset::Element { element_width, .. } => {
-            if info.array_dims.is_empty() {
-                return Err(crate::ir::verify::SirVerifyError::instruction(
-                    "MEMORY.ELEMENT_OFFSET_REQUIRES_UNPACKED_ARRAY",
-                    block,
-                    index,
-                    "SIROffset::Element used for a variable without unpacked dimensions",
-                ));
-            }
-            let Some(declared_element_width) = declared_element_width else {
-                return Err(crate::ir::verify::SirVerifyError::instruction(
-                    "MEMORY.UNPACKED_DECLARATION_HAS_ELEMENT_WIDTH",
-                    block,
-                    index,
-                    format!(
-                        "array dimensions {:?} do not divide declared width {}",
-                        info.array_dims, info.width
-                    ),
-                ));
-            };
-            if *element_width != declared_element_width {
-                return Err(crate::ir::verify::SirVerifyError::instruction(
-                    "MEMORY.ELEMENT_WIDTH_MATCHES_DECLARATION",
-                    block,
-                    index,
-                    format!(
-                        "SIR element width {element_width} does not match declared element width {declared_element_width}"
-                    ),
-                ));
-            }
-            let SIROffset::Element { bit_offset, .. } = offset else {
-                unreachable!()
-            };
-            if bit_offset
-                .checked_add(width)
-                .is_none_or(|end| end > declared_element_width)
-            {
-                return Err(crate::ir::verify::SirVerifyError::instruction(
-                    "MEMORY.ACCESS_STAYS_WITHIN_UNPACKED_ELEMENT",
-                    block,
-                    index,
-                    format!(
-                        "{operation} range [{bit_offset} +: {width}] exceeds unpacked element width {declared_element_width}"
-                    ),
-                ));
-            }
-        }
-        SIROffset::PackedElements {
-            bit_offset,
-            element_width,
-        } => {
-            let Some(declared_element_width) = declared_element_width else {
-                return Err(crate::ir::verify::SirVerifyError::instruction(
-                    "MEMORY.PACKED_ELEMENTS_REQUIRE_UNPACKED_ARRAY",
-                    block,
-                    index,
-                    format!("{operation} uses packed-elements addressing on a non-array variable"),
-                ));
-            };
-            let valid_range = *element_width == declared_element_width
-                && bit_offset.is_multiple_of(declared_element_width)
-                && width.is_multiple_of(declared_element_width)
-                && bit_offset
-                    .checked_add(width)
-                    .is_some_and(|end| end <= info.width);
-            if !valid_range {
-                return Err(crate::ir::verify::SirVerifyError::instruction(
-                    "MEMORY.PACKED_ELEMENTS_MATCH_DECLARATION",
-                    block,
-                    index,
-                    format!(
-                        "{operation} packed-elements range [{bit_offset} +: {width}] with element width {element_width} does not match declared element width {declared_element_width} and total width {}",
-                        info.width
-                    ),
-                ));
-            }
-        }
-        SIROffset::Static(start)
-            if !explicit_memory_copy
-                && let Some(element_width) = declared_element_width
-                && width != 0 =>
-        {
-            let Some(end) = start.checked_add(width) else {
-                return Err(crate::ir::verify::SirVerifyError::instruction(
-                    "MEMORY.ACCESS_STAYS_WITHIN_UNPACKED_ELEMENT",
-                    block,
-                    index,
-                    format!("{operation} range overflows usize"),
-                ));
-            };
-            if *start / element_width != end.saturating_sub(1) / element_width {
-                return Err(crate::ir::verify::SirVerifyError::instruction(
-                    "MEMORY.ACCESS_STAYS_WITHIN_UNPACKED_ELEMENT",
-                    block,
-                    index,
-                    format!(
-                        "{operation} at {addr:?} ({}) range [{start} +: {width}] crosses unpacked element width {element_width}; use an explicit array-copy operation for a multi-element transfer",
-                        program.get_path(&addr.absolute_addr()),
-                    ),
-                ));
-            }
-        }
-        SIROffset::Static(_) | SIROffset::Dynamic(_) => {}
-    }
-    Ok(())
+    celox_sir_opt::verify_memory_offset_contract(&program.design, eu)
 }
 
 fn verify_region_contract(
